Release v2.3.1
==============

New feature
-----------

* The "fake id0" feature was improved by Rémi Duraffort in order to
  support privileged write operations in read-only files/directories.
  Some package managers (Fedora, Debian, ...) relies on this special
  behavior::

      # ls -ld /usr/lib
      dr-xr-xr-x 22 root root 40960 Jan  2 11:19 /usr/lib/
      # install -v something.so /usr/lib/
      removed ‘/usr/lib/something.so‘
      ‘something.so‘ -> ‘/usr/lib/something.so‘

Fixes
-----

* Fix bindings to a guest path that contains a symbolic link.  For
  example when the given guest path ``/var/run/dbus`` is a symbolic
  link to ``/run/dbus``.

* Fix a memory corruption when accessing files in "/proc/self/"

Special thanks to Rémi Duraffort for the improved "fake id0" feature
and for the bug reports.

Conflicts:
	src/cli.h
	src/path/canon.c

+++ src/syscall/syscall.c
@@ -2,7 +2,7 @@
  *
  * This file is part of PRoot.
  *
- * Copyright (C) 2010, 2011, 2012 STMicroelectronics
+ * Copyright (C) 2013 STMicroelectronics
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License as
@@ -189,8 +189,8 @@ static int translate_syscall_enter(Tracee *tracee)
 			peek_reg(tracee, CURRENT, SYSARG_5), peek_reg(tracee, CURRENT, SYSARG_6),
 			peek_reg(tracee, CURRENT, STACK_POINTER));
 	else
-		VERBOSE(tracee, 2, "pid %d: syscall(%d)", tracee->pid,
-			(int)peek_reg(tracee, CURRENT, SYSARG_NUM));
+		VERBOSE(tracee, 2, "pid %d: syscall(%ld)", tracee->pid,
+			peek_reg(tracee, CURRENT, SYSARG_NUM));
 
 	status = notify_extensions(tracee, SYSCALL_ENTER_START, 0, 0);
 	if (status < 0)
